빅데이터 서버프로그램최종

+++ serverproject/src/main/java/site/metacoding/serverproject/domain/HospitalRepository.java
@@ -2,6 +2,6 @@ package site.metacoding.serverproject.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HospitalRepository extends JpaRepository<Hospital, Integer> {
+public interface HospitalRepository extends JpaRepository<Hospital, Long> {
 
 }
